touch up sequence diagram
