+++ docs/puml/sequence_diagrams/gym_station_sequence.puml
@@ -1,29 +1,20 @@
 @startuml
 mainframe sd parseGymStationInput
 actor User
-participant "handler:Handler" as Handler
 participant "parser:Parser" as Parser
-participant "validation:Validation" as Validation
 participant "newGym:Gym" as Gym
 participant "output:Output" as Output
+participant "gymStation:GymStation" as GymStation
 
 activate Parser  #FFBBBB
+
 Parser -> Parser : parseGymStationInput()
     activate Parser #DarkSalmon
         loop until all gym stations recorded
-            Parser -> Output : printGymStationPrompt()
-            activate Output #FFBBBB
-            Output --> Parser : 
-            deactivate Output #DarkSalmon
 
-            Parser -> Handler: in.nextLine()
-            activate Handler #FFBBBB
-            Handler -> User:
-            User --> Handler: user input
-            Handler --> Parser: user input
-            deactivate Handler #FFBBBB
+            User -> Parser: user input
             Parser -> Parser: splitGymStationInput()
-            activate Parser #DarkSalmon
+            activate Parser #FFBBBB
             Parser --> Parser:
 
             deactivate Parser #DarkSalmon
@@ -33,15 +24,40 @@ Parser -> Parser : parseGymStationInput()
 
             activate Gym #FFBBBB
 
-            Gym -> Gym: validateParameters()
-            activate Gym #Salmon
-            Gym --> Gym: 
-            deactivate Gym #Salmon
-
             Gym -> Gym : appendIntoStations()
-            activate Gym #Salmon
+            activate Gym #DarkSalmon
+                Gym -> GymStation **: new
+                activate GymStation #FFBBBB
+
+                GymStation -> GymStation: checkInput
+                note right: Consist of multiple \n validation checks
+                activate GymStation #Salmon
+                GymStation --> GymStation:
+                deactivate GymStation #Salmon
+
+                GymStation -> GymStation : processSets()
+                activate GymStation #DarkSalmon
+
+                    loop until all sets are recorded
+                        GymStation -> GymSet **: new 
+                        activate GymSet #FFBBBB
+                        GymSet --> GymStation:
+                        deactivate GymSet #FFBBBB
+                        destroy GymSet
+
+                    end
+
+
+                GymStation --> GymStation
+                deactivate GymStation #DarkSalmon
+
+
+                GymStation --> Gym: 
+                deactivate GymStation #FFBBBB
+                destroy GymStation
+
             Gym --> Gym : 
-            deactivate Gym #Salmon
+            deactivate Gym #DarkSalmon
             Gym --> Parser:
             deactivate Gym #FFBBBB
         end
